requirements.txt
alembic
Flask
Flask_Migrate
flask_sqlalchemy
PyJWT
PyJWT
SQLAlchemy
Werkzeug